bushtree/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
boto3==1.35.19
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.4.1
cron-descriptor==1.4.0
cryptography==41.0.7
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.1
django-autocomplete-light==3.9.7
django-celery-beat==2.5.0
django-cors-headers==4.3.1
django-filter==23.5
django-import-export==3.3.5
django-oauth-toolkit==2.3.0
django-redis==5.4.0
django-storages==1.14.4
django-rest-framework-social-oauth2==1.2.0
django-templated-mail==1.1.1
django-timezone-field==6.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
djongo==1.3.6
dnspython==2.4.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
eventlet==0.34.3
exceptiongroup==1.2.0
flake8==6.1.0
greenlet==3.0.3
gunicorn==21.2.0
idna==3.6
inflection==0.5.1
iniconfig==2.0.0
isort==5.13.2
jwcrypto==1.5.6
jupyterlab==4.2.5
kombu==5.3.4
MarkupPy==1.14
mccabe==0.7.0
numpy==1.26.4
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
packaging==23.2
pandas==2.2.2
pdfrw==0.4
pillow==10.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
pytest==8.0.1
pytest-django==4.8.0
python-crontab==3.0.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
redis==5.0.1
reportlab==4.0.8
requests==2.31.0
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.1
SQLAlchemy==2.0.30
sqlparse==0.2.4
tablib==3.5.0
tomli==2.0.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
vine==5.1.0
wcwidth==0.2.12
xlrd==2.0.1
xlwt==1.3.0
serializers==0.2.4
future==1.0.0

scikit-learn